fixup! Add a job that runs as close to midnight as possible
